SQUASHME! [VPG]: drm/i915: Decrease TDR hang recovery latency
When scheduling/rescheduling hang detection we round up to an integral number
of seconds. This is not so good when rounding from just above one second since
it then becomes exactly two seconds in between hang checks. The resulting hang
detection latency goes from 3*1=3 seconds to 3*2=6 seconds, which is way too
long for certain applications such as userland HW compositor, which sometimes
results in broken flips and frozen display. In this change we remove this
rounding.

Another change is that we remove the seqno pre-check in queue_hangcheck that
adds another stage before the real hang check. Basically, this adds one hang
check period on top of what is already specified, which is not necessary. The
reason why it was originally added was to reduce the amounts of false positives
but as it turns out we've had more problems with hang recovery latency and not
so much problems with false positives so there's no reason for it to still be
in there.

    SQUASHME! - This patch should be squashed into the following patch:
    	Author: Tomas Elf <tomas.elf@intel.com>
    	Change-Id: I45de965b4a81d93c135a74248eb22a8a5c816e9f
    	FOR_UPSTREAM [VPG]: drm/i915: Added gen8 support for Timeout Detection Recovery for engine hangs

Change-Id: I192729693a8541e152fb3dacf793e0f0f69f505f
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-14282
Signed-off-by: Tomas Elf <tomas.elf@intel.com>

## Don't forget to add a JIRA URL to the Tracked-On metadata!
## https://wiki.ith.intel.com/display/CACTUS/Patch+requirements+check
## url should be in the form:
## Tracked-On: https://jira01.devtools.intel.com/browse/<PROJ>-<NUMB>
## Several Tracked-On lines are allowed
## you can fully prefill your Tracked-On commits using environment variable:
## export CURRENT_TRACKED_ON=https://jira01.devtools.intel.com/browse/<PROJ>-<NUMB>
##