REVERTME [VPG]: drm/i915: Dedicated retire work workqueue.
Created a dedicated, high-priority workqueue for retire work handler. The
reason for this is that in customer issue IMINAN-42241 the Linux kernel queues
the retire work work item in time but blocks the activation and subsequent
execution of this deferred activity for more than a minute in the face of a GPU
engine hang, which in turn blocks the scheduling of the hang checker (that is
piggy-backed on top of the retire work handler). This means that the system is
hung for more than a minute without as much as an attempt to detect the hang,
which causes Android userland to start killing off processes. The reason for
the retire work handler being blocked is that the shared i915 workqueue that is
normally used for the retire work handler gets congested during high-load
scenarios (such as video playback on 3 displays, out of which one is over WIDI
with HDCP). It's not acceptable that hang detection is blocked for that long
and therefore we need to do something to ensure that the Linux kernel gets
around to scheduling the retire work handler work item quicker. By creating a
dedicated, high-priority workqueue we can make it less likely for this problem
to arise, but obviously since this is not a functional bug but rather a flaw
that is exposed in high-load scenarios it's conceivable that another scenario
with significantly higher load could cause this problem to resurface. However,
at some point we have to realise that the system can only operate correctly
under so much load (it's not a real-time system after all) and the question is
really where we draw the line for the maximum tolerated load.

It should be noted that the repro steps in IMINAN-42241 are not part of a POR
use-case since it requires you to modify /system/build.prop to enable
three-display clone mode. So you might make the case that this kind of load is
not acceptable to begin with.

This is REVERTME because the TDR version currently being upstreamed does not
rely on the retire work handler to handle hang check scheduling. Once we
upstream TDR and forklift it back into Android the retire work handler will
become irrelevant to the hang checker. But until then TDR needs to make sure
the retire work handler is scheduled flawlessly.

Change-Id: I5dc3ec4e04c698ddfdebcfc25e3265cc32f8d103
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-40580
Signed-off-by: Tomas Elf <tomas.elf@intel.com>
Signed-off-by: Antonio Argenziano <antonio.argenziano@intel.com>
Reviewed-on: https://android.intel.com:443/554437